refactor: split remaining CODE-4 god-files
